+++ src/OperatingSystem/Implementation.php
@@ -1,0 +1,35 @@
+<?php
+declare(strict_types = 1);
+
+namespace Innmind\OperatingSystem\OperatingSystem;
+
+use Innmind\OperatingSystem\{
+    Filesystem,
+    Ports,
+    Sockets,
+    Remote,
+    CurrentProcess,
+    Config,
+};
+use Innmind\Server\Status\Server as ServerStatus;
+use Innmind\Server\Control\Server as ServerControl;
+use Innmind\TimeContinuum\Clock;
+
+interface Implementation
+{
+    /**
+     * This method allows to change the underlying OS implementation while being
+     * able to keep any decorators on top of it.
+     *
+     * @param callable(self, Config): self $map
+     */
+    public function map(callable $map): self;
+    public function clock(): Clock;
+    public function filesystem(): Filesystem;
+    public function status(): ServerStatus;
+    public function control(): ServerControl;
+    public function ports(): Ports;
+    public function sockets(): Sockets;
+    public function remote(): Remote;
+    public function process(): CurrentProcess;
+}